added norm integration into travis
added a make norm rule which is called in make test.
the norm rule runs a script which runs both norminette and norminette+.
if either has errors, the script exits with -1, causing the makefile to error.
This causes travis to fail before even running any other tests.
(just like moulinette!)