Feature/build dispatcher (#39)
* added new functions for dispatcher

Added dispatcher .c file, and some other files to create working proof of concept of how the dispatcher and parsing will work.

* modified ft_print.c for dispatcher

modified ft_printf.c for proof of concept of the dispatcher

* modified ft_printf.h and Sourcesfor dispatcher

updated the ft_printf header file for the dispatcher proof of concept. should possibly be split up in the future to a dispatcher.h
Also added new functions to Sources file

* added basic tests for dispatcher proof of concept

added basic tests to cover the proof of concept functionality of the dispatcher.
for some reason coverage doesn't seem to detect a part of one function, even though I'm 90% sure it is covered...
wil ltry to fix later.

* added c files for strchar functions

added 2 functions for adding a char to the end of a string.

* updated support files for strchar functions

added the strchar functions to the libft header
and the covsources file.

* added tests for strchar functions

added tests for strcharjoin and strcharexpand.
also updated testsources to include these.

* added slightly better placeholder formatters

improved ft_formatchar and ft_formatstring,
and addedd ft_formatunknown, which handles unknown conversions.

* added ft_process_conversion

ft_process_conversion is a central function handling the entire conversion.

* slightly updated ft_init_info

still not done, but slightly improved.
no using the enums for options.

* fixes in strchar functions

some errors in ft_strchar functions fixed.
working properly now.

* update of ft_printf.c

updated the ft_printf functions to call ft_proceess_conversion.

* fixed missing backslash in covsources

* enums and new functions added to printf.h

* fixed incorrect test

this test assumed error on unknown specifier.
now assumes correct handling of the unknown specifier.

* moved init of writer to ft_printf.c

since this should stay the same, it is initialized once in ft_printf.

* changed the dispatcher index to unsigned

since this is an index, we want it to explicitely be unsigned.

* fixed space print in formatchar and string

it added the space at the end instead front

* added more tests

got as much coverage as possible.