Feature/formatint (#66)
* add formatint and handle negative int

* started testing

* int tests

* last tests added

* restored updated date as a test

* pretty please merge

* format int changed for new stuff

* small fixes

* changed flag order in Makefile

for better compatibility with other systems.
(on Debian in WSL on Windows it won't compile if the flags are at the end)

* small fix for include paths

* added i to dispatcher

* restructured formatint funcs for lenmod

ft_formatint is now updated for the new workflow with ft_formatdecimal,
This is so we can integrate lenmod. ft_handle_negint is now ft_format_negint.
That function is also updated, it is now an alternative to ft_formatint instead of a part of.

* added ft_format_decimal and PLACEHOLDERS

added ft_formatdecimal which will dispatch `i` and `d` conversions,
depending on the length modifier provided.
also added placeholders that need to be build for lenmod conversions.

* small change to support larger numbers

* added ft_write_flags

this function will write flags when needed.

* modified header for new and modified functions.

* added and updated tests for int

added some tests, mostly integration tests calling the ft_printf function.

* added new functions to sources file

* small fixes from merge

* added negative test with - flag

* made ft_write_flags modify len properly.

* stash pop

* added tests for int with + and large width

this test was missing and failing, so added them.

* changed the way we handle final processing

made a subfunction that controls the order of the final processing.
This also fixes an issue with a large width and the + flag.
Also integrated negint as static function in file of formatint.

* added more complete version of formatbyteint

no precision yet, but more complete.

* small change to gitignore

* added more final version of formatshort

* removed formatlong

since on our system formatlong is the same as formatlonglong,
I removed it. On other systems it would be the same as formatint, so it would be easy to switch.

* added formatlonglong and tests for longlong and long